Switch to parallel test suite
