+++ systemtest/src/test/java/io/strimzi/systemtest/specific/RackAwarenessST.java
@@ -1,0 +1,223 @@
+/*
+ * Copyright Strimzi authors.
+ * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
+ */
+package io.strimzi.systemtest.specific;
+
+import io.fabric8.kubernetes.api.model.Affinity;
+import io.fabric8.kubernetes.api.model.NodeAffinity;
+import io.fabric8.kubernetes.api.model.NodeSelectorRequirement;
+import io.fabric8.kubernetes.api.model.Pod;
+import io.fabric8.kubernetes.api.model.PodAffinityTerm;
+import io.fabric8.kubernetes.api.model.PodStatus;
+import io.strimzi.api.kafka.model.KafkaConnect;
+import io.strimzi.api.kafka.model.KafkaConnectResources;
+import io.strimzi.api.kafka.model.KafkaMirrorMaker2Resources;
+import io.strimzi.api.kafka.model.KafkaResources;
+import io.strimzi.api.kafka.model.Rack;
+import io.strimzi.operator.common.model.Labels;
+import io.strimzi.systemtest.AbstractST;
+import io.strimzi.systemtest.Constants;
+import static io.strimzi.systemtest.Constants.CO_OPERATION_TIMEOUT_SHORT;
+import static io.strimzi.systemtest.Constants.REGRESSION;
+import io.strimzi.systemtest.Environment;
+import io.strimzi.systemtest.annotations.ParallelNamespaceTest;
+import io.strimzi.systemtest.annotations.ParallelSuite;
+import static io.strimzi.systemtest.resources.ResourceManager.kubeClient;
+import io.strimzi.systemtest.resources.crd.KafkaConnectResource;
+import io.strimzi.systemtest.storage.TestStorage;
+import io.strimzi.systemtest.templates.crd.KafkaConnectTemplates;
+import io.strimzi.systemtest.templates.crd.KafkaMirrorMaker2Templates;
+import io.strimzi.systemtest.templates.crd.KafkaTemplates;
+import io.strimzi.systemtest.utils.StUtils;
+import io.strimzi.systemtest.utils.kafkaUtils.KafkaConnectUtils;
+import io.strimzi.systemtest.utils.kubeUtils.objects.PodUtils;
+import static io.strimzi.test.k8s.KubeClusterResource.cmdKubeClient;
+import java.util.List;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.hasEntry;
+import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assumptions.assumeFalse;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.extension.ExtensionContext;
+
+@Tag(REGRESSION)
+@ParallelSuite
+class RackAwarenessST extends AbstractST {
+    private static final Logger LOGGER = LogManager.getLogger(RackAwarenessST.class);
+    private static final String TOPOLOGY_KEY = "kubernetes.io/hostname";
+
+    @ParallelNamespaceTest
+    void testKafkaRackAwareness(ExtensionContext extensionContext) {
+        TestStorage storage = storageMap.get(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+
+        resourceManager.createResource(extensionContext, KafkaTemplates.kafkaEphemeral(clusterName, 1, 1)
+                .editSpec()
+                    .editKafka()
+                        .withNewRack(TOPOLOGY_KEY)
+                        .addToConfig("replica.selector.class", "org.apache.kafka.common.replica.RackAwareReplicaSelector")
+                    .endKafka()
+                .endSpec()
+                .build());
+
+        LOGGER.info("Kafka cluster deployed successfully");
+        String ssName = KafkaResources.kafkaStatefulSetName(clusterName);
+        String podName = PodUtils.getPodNameByPrefix(namespace, ssName);
+        Pod pod = kubeClient().getPod(namespace, podName);
+
+        // check that spec matches the actual pod configuration
+        Affinity specAffinity = StUtils.getStatefulSetOrStrimziPodSetAffinity(namespace, KafkaResources.kafkaStatefulSetName(clusterName));
+        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
+        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        assertThat(podNodeRequirement, is(specNodeRequirement));
+        assertThat(specNodeRequirement.getKey(), is(TOPOLOGY_KEY));
+        assertThat(specNodeRequirement.getOperator(), is("Exists"));
+
+        PodAffinityTerm specPodAntiAffinityTerm = specAffinity.getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
+        PodAffinityTerm podAntiAffinityTerm = pod.getSpec().getAffinity().getPodAntiAffinity().getPreferredDuringSchedulingIgnoredDuringExecution().get(0).getPodAffinityTerm();
+        assertThat(podAntiAffinityTerm, is(specPodAntiAffinityTerm));
+        assertThat(specPodAntiAffinityTerm.getTopologyKey(), is(TOPOLOGY_KEY));
+        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/cluster", clusterName));
+        assertThat(specPodAntiAffinityTerm.getLabelSelector().getMatchLabels(), hasEntry("strimzi.io/name", KafkaResources.kafkaStatefulSetName(clusterName)));
+
+        // check Kafka rack awareness configuration
+        String podNodeName = pod.getSpec().getNodeName();
+        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
+
+        String rackIdOut = cmdKubeClient(namespace).execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /opt/kafka/init/rack.id").out().trim();
+        String brokerRackOut = cmdKubeClient(namespace).execInPod(KafkaResources.kafkaPodName(clusterName, 0), "/bin/bash", "-c", "cat /tmp/strimzi.properties | grep broker.rack").out().trim();
+        assertThat(rackIdOut.trim(), is(hostname));
+        assertThat(brokerRackOut.contains("broker.rack=" + hostname), is(true));
+    }
+
+    @ParallelNamespaceTest
+    void testConnectRackAwareness(ExtensionContext extensionContext) {
+        TestStorage storage = storageMap.get(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+
+        resourceManager.createResource(extensionContext, KafkaTemplates.kafkaEphemeral(clusterName, 1, 1).build());
+        resourceManager.createResource(extensionContext, KafkaConnectTemplates.kafkaConnect(extensionContext, clusterName, 1)
+                .editSpec()
+                    .withNewRack(TOPOLOGY_KEY)
+                .endSpec()
+                .build());
+
+        LOGGER.info("Connect cluster deployed successfully");
+        String deployName = KafkaConnectResources.deploymentName(clusterName);
+        String podName = PodUtils.getPodNameByPrefix(namespace, deployName);
+        Pod pod = kubeClient().getPod(namespace, podName);
+
+        // check that spec matches the actual pod configuration
+        Affinity specAffinity = kubeClient().getDeployment(namespace, deployName).getSpec().getTemplate().getSpec().getAffinity();
+        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
+        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        assertThat(podNodeRequirement, is(specNodeRequirement));
+        assertThat(podNodeRequirement.getKey(), is(TOPOLOGY_KEY));
+        assertThat(podNodeRequirement.getOperator(), is("Exists"));
+
+        // check Kafka client rack awareness configuration
+        String podNodeName = pod.getSpec().getNodeName();
+        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
+        String commandOut = cmdKubeClient(namespace).execInPod(podName, "/bin/bash", "-c", "cat /tmp/strimzi-connect.properties | grep consumer.client.rack").out().trim();
+        assertThat(commandOut.equals("consumer.client.rack=" + hostname), is(true));
+    }
+
+    @ParallelNamespaceTest
+    void testConnectRackAwarenessWithInvalidTopologyKey(ExtensionContext extensionContext) {
+        TestStorage storage = storageMap.get(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+        String invalidTopologyKey = "invalid-topology-key";
+
+        resourceManager.createResource(extensionContext, KafkaTemplates.kafkaEphemeral(clusterName, 1, 1).build());
+
+        LOGGER.info("Deploy KafkaConnect with invalid topology key: {}", invalidTopologyKey);
+        resourceManager.createResource(extensionContext, false, KafkaConnectTemplates.kafkaConnect(extensionContext, clusterName, 1)
+                .editSpec()
+                    .withNewRack(invalidTopologyKey)
+                .endSpec()
+                .build());
+
+        LOGGER.info("Waiting for ClusterOperator to get timeout operation of incorrect KafkaConnect resource");
+        KafkaConnectUtils.waitForKafkaConnectCondition("TimeoutException", "NotReady", namespace, clusterName);
+
+        List<String> connectNotReadyPods = kubeClient().listPodNames(namespace, clusterName, Labels.STRIMZI_KIND_LABEL, KafkaConnect.RESOURCE_KIND);
+        PodStatus kcWrongStatus = kubeClient().getPod(namespace, connectNotReadyPods.get(0)).getStatus();
+        assertThat("Unschedulable", is(kcWrongStatus.getConditions().get(0).getReason()));
+        assertThat("PodScheduled", is(kcWrongStatus.getConditions().get(0).getType()));
+
+        LOGGER.info("Fix KafkaConnect resource with valid topology key: {}", TOPOLOGY_KEY);
+        KafkaConnectResource.replaceKafkaConnectResourceInSpecificNamespace(clusterName, kc -> kc.getSpec().setRack(new Rack(TOPOLOGY_KEY)), namespace);
+        KafkaConnectUtils.waitForConnectReady(namespace, clusterName);
+
+        KafkaConnect kc = KafkaConnectResource.kafkaConnectClient().inNamespace(namespace).withName(clusterName).get();
+        assertThat(kc.getSpec().getRack().getTopologyKey(), is(TOPOLOGY_KEY));
+    }
+
+    @ParallelNamespaceTest
+    void testMirrorMaker2RackAwareness(ExtensionContext extensionContext) {
+        TestStorage storage = storageMap.get(extensionContext);
+        String namespace = storage.getNamespaceName();
+        String clusterName = storage.getClusterName();
+        String kafkaClusterSourceName = clusterName + "-source";
+        String kafkaClusterTargetName = clusterName + "-target";
+
+        resourceManager.createResource(extensionContext,
+                KafkaTemplates.kafkaEphemeral(kafkaClusterSourceName, 1, 1).build());
+        resourceManager.createResource(extensionContext,
+                KafkaTemplates.kafkaEphemeral(kafkaClusterTargetName, 1, 1).build());
+
+        resourceManager.createResource(extensionContext,
+                KafkaMirrorMaker2Templates.kafkaMirrorMaker2(clusterName, kafkaClusterTargetName, kafkaClusterSourceName, 1, false)
+                        .editSpec()
+                            .withNewRack(TOPOLOGY_KEY)
+                        .endSpec()
+                        .build());
+
+        LOGGER.info("MirrorMaker2 cluster deployed successfully");
+        String deployName = KafkaMirrorMaker2Resources.deploymentName(clusterName);
+        String podName = PodUtils.getPodNameByPrefix(namespace, deployName);
+        Pod pod = kubeClient().getPod(namespace, podName);
+
+        // check that spec matches the actual pod configuration
+        Affinity specAffinity = kubeClient().getDeployment(namespace, deployName).getSpec().getTemplate().getSpec().getAffinity();
+        NodeSelectorRequirement specNodeRequirement = specAffinity.getNodeAffinity().getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        NodeAffinity podAffinity = pod.getSpec().getAffinity().getNodeAffinity();
+        NodeSelectorRequirement podNodeRequirement = podAffinity.getRequiredDuringSchedulingIgnoredDuringExecution().getNodeSelectorTerms().get(0).getMatchExpressions().get(0);
+        assertThat(podNodeRequirement, is(specNodeRequirement));
+        assertThat(podNodeRequirement.getKey(), is(TOPOLOGY_KEY));
+        assertThat(podNodeRequirement.getOperator(), is("Exists"));
+
+        // check Kafka client rack awareness configuration
+        String podNodeName = pod.getSpec().getNodeName();
+        String hostname = podNodeName.contains(".") ? podNodeName.substring(0, podNodeName.indexOf(".")) : podNodeName;
+        String commandOut = cmdKubeClient(namespace).execInPod(podName, "/bin/bash", "-c", "cat /tmp/strimzi-connect.properties | grep consumer.client.rack").out().trim();
+        assertThat(commandOut.equals("consumer.client.rack=" + hostname), is(true));
+    }
+
+    @BeforeEach
+    void createTestResources(ExtensionContext extensionContext) {
+        storageMap.put(extensionContext, new TestStorage(extensionContext));
+    }
+
+    @BeforeAll
+    void setUp() {
+        assumeFalse(Environment.isNamespaceRbacScope());
+        clusterOperator.unInstall();
+        clusterOperator = clusterOperator
+                .defaultInstallation()
+                .withOperationTimeout(CO_OPERATION_TIMEOUT_SHORT)
+                .withReconciliationInterval(Constants.RECONCILIATION_INTERVAL)
+                .createInstallation()
+                .runInstallation();
+    }
+}
