+++ test/src/main/java/io/strimzi/test/k8s/KubeClient.java
@@ -328,7 +328,11 @@ public class KubeClient {
      * Gets pod Uid
      */
     public String getPodUid(String name) {
-        return client.pods().inNamespace(getNamespace()).withName(name).get().getMetadata().getUid();
+        return getPodUid(getNamespace(), name);
+    }
+
+    public String getPodUid(String namespace, String name) {
+        return client.pods().inNamespace(namespace).withName(name).get().getMetadata().getUid();
     }
 
     /**
@@ -431,7 +435,6 @@ public class KubeClient {
     /**
      * Gets deployment
      */
-
     public Deployment getDeployment(String namespaceName, String deploymentName) {
         return client.apps().deployments().inNamespace(namespaceName).withName(deploymentName).get();
     }
